requirements.txt
# Core PyTorch dependencies
torch>=1.8.1
torchvision>=0.9.1
torchaudio>=0.8.1

# Project-specific dependencies
pytorch-lightning>=1.3.3
mlflow>=1.17.0
gitpython>=3.1.14
performer-pytorch

# Core data science and ML libraries
numpy>=1.19.2
pandas>=1.1.5
scipy>=1.5.4
scikit-learn>=0.24.0

# Image processing
pillow>=8.2.0
opencv-python>=4.5.2

# Visualization and utilities
matplotlib>=3.3.4
tqdm>=4.60.0

# NLP and transformers
transformers>=4.6.1
tokenizers>=0.10.3
ftfy>=6.0.1
regex>=2021.4.4

# Specialized libraries from your environment
einops>=0.3.0
vector-quantize-pytorch>=0.1.0
timm>=0.4.5
tensorboard>=2.4.1
tensorboardx>=1.8
trimesh>=3.9.16
pynrrd>=0.4.2
axial-positional-embedding>=0.2.1

fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5  # For file uploads